+++ darts/ad/detectors/threshold_detector.py
@@ -12,9 +12,9 @@ from typing import Union
 
 import numpy as np
 
+from darts import TimeSeries
 from darts.ad.detectors.detectors import Detector, _BoundedDetectorMixin
 from darts.logging import get_logger, raise_log
-from darts.timeseries import TimeSeries
 
 logger = get_logger(__name__)
