+++ darts/dataprocessing/transformers/boxcox.py
@@ -11,6 +11,7 @@ import pandas as pd
 from scipy.special import inv_boxcox
 from scipy.stats import boxcox, boxcox_normmax
 
+from darts import TimeSeries
 from darts.dataprocessing.transformers.fittable_data_transformer import (
     FittableDataTransformer,
 )
@@ -18,7 +19,6 @@ from darts.dataprocessing.transformers.invertible_data_transformer import (
     InvertibleDataTransformer,
 )
 from darts.logging import get_logger, raise_if
-from darts.timeseries import TimeSeries
 
 logger = get_logger(__name__)
