+++ darts/dataprocessing/transformers/mappers.py
@@ -9,12 +9,12 @@ from typing import Any, Callable, Union
 import numpy as np
 import pandas as pd
 
+from darts import TimeSeries
 from darts.dataprocessing.transformers.base_data_transformer import BaseDataTransformer
 from darts.dataprocessing.transformers.invertible_data_transformer import (
     InvertibleDataTransformer,
 )
 from darts.logging import get_logger
-from darts.timeseries import TimeSeries
 
 logger = get_logger(__name__)
