+++ darts/dataprocessing/transformers/scaler.py
@@ -10,6 +10,7 @@ from typing import Any, Union
 import numpy as np
 from sklearn.preprocessing import MinMaxScaler
 
+from darts import TimeSeries
 from darts.dataprocessing.transformers.fittable_data_transformer import (
     FittableDataTransformer,
 )
@@ -17,7 +18,6 @@ from darts.dataprocessing.transformers.invertible_data_transformer import (
     InvertibleDataTransformer,
 )
 from darts.logging import get_logger, raise_log
-from darts.timeseries import TimeSeries
 
 logger = get_logger(__name__)
