+++ darts/models/filtering/filtering_model.py
@@ -7,8 +7,8 @@ returns a `TimeSeries` that is a filtered version of `series`.
 
 from abc import ABC, abstractmethod
 
+from darts import TimeSeries
 from darts.logging import get_logger, raise_if_not
-from darts.timeseries import TimeSeries
 
 logger = get_logger(__name__)
