+++ darts/models/filtering/gaussian_process_filter.py
@@ -9,8 +9,8 @@ import numpy as np
 from sklearn.gaussian_process import GaussianProcessRegressor
 from sklearn.gaussian_process.kernels import Kernel
 
+from darts import TimeSeries
 from darts.models.filtering.filtering_model import FilteringModel
-from darts.timeseries import TimeSeries
 
 
 class GaussianProcessFilter(FilteringModel):
