+++ darts/models/filtering/moving_average_filter.py
@@ -3,8 +3,8 @@ Moving Average
 --------------
 """
 
+from darts import TimeSeries
 from darts.models.filtering.filtering_model import FilteringModel
-from darts.timeseries import TimeSeries
 
 
 class MovingAverageFilter(FilteringModel):
