+++ darts/models/forecasting/catboost_model.py
@@ -14,12 +14,12 @@ import numpy as np
 import pandas as pd
 from catboost import CatBoostRegressor, Pool
 
+from darts import TimeSeries
 from darts.logging import get_logger
 from darts.models.forecasting.sklearn_model import (
     SKLearnModelWithCategoricalCovariates,
     _QuantileModelContainer,
 )
-from darts.timeseries import TimeSeries
 from darts.utils.likelihood_models.sklearn import (
     QuantileRegression,
     _check_likelihood,
