+++ darts/models/forecasting/fft.py
@@ -9,9 +9,9 @@ import numpy as np
 import pandas as pd
 from statsmodels.tsa.stattools import acf
 
+from darts import TimeSeries
 from darts.logging import get_logger
 from darts.models.forecasting.forecasting_model import LocalForecastingModel
-from darts.timeseries import TimeSeries
 from darts.utils.missing_values import fill_missing_values
 
 logger = get_logger(__name__)
