+++ darts/models/forecasting/kalman_forecaster.py
@@ -15,12 +15,12 @@ from typing import Optional
 import numpy as np
 from nfoursid.kalman import Kalman
 
+from darts import TimeSeries
 from darts.logging import get_logger
 from darts.models.filtering.kalman_filter import KalmanFilter
 from darts.models.forecasting.forecasting_model import (
     TransferableFutureCovariatesLocalForecastingModel,
 )
-from darts.timeseries import TimeSeries
 
 logger = get_logger(__name__)
