+++ darts/models/forecasting/lgbm.py
@@ -15,6 +15,7 @@ from typing import Optional, Union
 
 import lightgbm as lgb
 
+from darts import TimeSeries
 from darts.logging import get_logger
 from darts.models.forecasting.sklearn_model import (
     FUTURE_LAGS_TYPE,
@@ -22,7 +23,6 @@ from darts.models.forecasting.sklearn_model import (
     SKLearnModelWithCategoricalCovariates,
     _QuantileModelContainer,
 )
-from darts.timeseries import TimeSeries
 from darts.utils.likelihood_models.sklearn import (
     QuantileRegression,
     _check_likelihood,
