+++ darts/models/forecasting/prophet_model.py
@@ -12,11 +12,11 @@ import numpy as np
 import pandas as pd
 import prophet
 
+from darts import TimeSeries
 from darts.logging import execute_and_suppress_output, get_logger, raise_if, raise_log
 from darts.models.forecasting.forecasting_model import (
     FutureCovariatesLocalForecastingModel,
 )
-from darts.timeseries import TimeSeries
 
 logger = get_logger(__name__)
 logger.level = logging.WARNING  # set to warning to suppress prophet logs
