+++ darts/models/forecasting/theta.py
@@ -9,9 +9,9 @@ from typing import Optional
 import numpy as np
 import statsmodels.tsa.holtwinters as hw
 
+from darts import TimeSeries
 from darts.logging import get_logger, raise_if_not, raise_log
 from darts.models.forecasting.forecasting_model import LocalForecastingModel
-from darts.timeseries import TimeSeries
 from darts.utils.statistics import (
     check_seasonality,
     extract_trend_and_seasonality,
