+++ darts/utils/historical_forecasts/optimized_historical_forecasts_torch.py
@@ -5,8 +5,8 @@ from typing import Literal, Optional, Union
 import numpy as np
 import pandas as pd
 
+from darts import TimeSeries
 from darts.logging import get_logger
-from darts.timeseries import TimeSeries
 from darts.utils.historical_forecasts.utils import (
     _get_historical_forecast_boundaries,
     _process_predict_start_points_bounds,
